# docs/openapi/critical-endpoints.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Tamil Platform Critical API Contracts",
    "version": "1.0.0",
    "description": "Critical endpoint OpenAPI contract used for launch-readiness response governance."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "tags": ["Auth"],
        "operationId": "getProfile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerializedUser"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Auth"],
        "operationId": "updateProfile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerializedUser"
                }
              }
            }
          }
        }
      }
    },
    "/auth/adaptive-profile": {
      "get": {
        "tags": ["Auth"],
        "operationId": "getAdaptiveProfileInsights",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Adaptive profile insights with category recommendations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptiveProfileInsightsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgotpassword": {
      "post": {
        "tags": ["Auth"],
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic password reset response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/resetpassword/{resettoken}": {
      "put": {
        "tags": ["Auth"],
        "operationId": "resetPassword",
        "parameters": [
          {
            "in": "path",
            "name": "resettoken",
            "required": true,
            "schema": { "type": "string", "minLength": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["password"],
                "properties": {
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Badge": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "icon"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "icon": { "type": "string" },
          "earnedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SerializedUser": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "email",
          "skill_score",
          "level",
          "lessons_completed",
          "unlockedStages",
          "masteryPassedStages",
          "xp",
          "totalXP",
          "current_streak",
          "longest_streak",
          "badges",
          "avatarId",
          "adaptivePreferences",
          "adaptiveProfile",
          "role",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "skill_score": { "type": "number" },
          "level": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          },
          "lessons_completed": { "type": "integer", "minimum": 0 },
          "unlockedStages": {
            "type": "array",
            "items": { "type": "integer", "minimum": 1, "maximum": 10 }
          },
          "masteryPassedStages": {
            "type": "array",
            "items": { "type": "integer", "minimum": 1, "maximum": 10 }
          },
          "xp": { "type": "number", "minimum": 0 },
          "totalXP": { "type": "number", "minimum": 0 },
          "current_streak": { "type": "number", "minimum": 0 },
          "longest_streak": { "type": "number", "minimum": 0 },
          "badges": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Badge" }
          },
          "avatarId": { "type": "string" },
          "adaptivePreferences": { "$ref": "#/components/schemas/AdaptivePreferences" },
          "adaptiveProfile": { "$ref": "#/components/schemas/AdaptiveProfile" },
          "role": { "type": "string", "enum": ["user", "admin"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AdaptivePreferences": {
        "type": "object",
        "additionalProperties": false,
        "required": ["modePreference", "immersiveModeDefault"],
        "properties": {
          "modePreference": {
            "type": "string",
            "enum": ["auto", "support", "balanced", "challenge"]
          },
          "immersiveModeDefault": { "type": "boolean" }
        }
      },
      "AdaptiveProfile": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "recommendedMode",
          "recommendedDifficulty",
          "supportNeed",
          "challengeReadiness",
          "stabilityScore",
          "confidenceScore",
          "lastUpdatedAt"
        ],
        "properties": {
          "recommendedMode": {
            "type": "string",
            "enum": ["support", "balanced", "challenge"]
          },
          "recommendedDifficulty": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          },
          "supportNeed": { "type": "number", "minimum": 0, "maximum": 100 },
          "challengeReadiness": { "type": "number", "minimum": 0, "maximum": 100 },
          "stabilityScore": { "type": "number", "minimum": 0, "maximum": 100 },
          "confidenceScore": { "type": "number", "minimum": 0, "maximum": 100 },
          "lastUpdatedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "AdaptiveCategoryRecommendation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "category",
          "recommendedMode",
          "recommendedDifficulty",
          "supportNeed",
          "challengeReadiness",
          "confidenceScore",
          "attemptCount",
          "successRate",
          "avgErrors",
          "avgHints",
          "avgRetries",
          "priorityScore",
          "reason"
        ],
        "properties": {
          "category": {
            "type": "string",
            "enum": ["uyir", "mei", "uyir-mei", "grammar", "sentences"]
          },
          "recommendedMode": {
            "type": "string",
            "enum": ["support", "balanced", "challenge"]
          },
          "recommendedDifficulty": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          },
          "supportNeed": { "type": "number", "minimum": 0, "maximum": 100 },
          "challengeReadiness": { "type": "number", "minimum": 0, "maximum": 100 },
          "confidenceScore": { "type": "number", "minimum": 0, "maximum": 100 },
          "attemptCount": { "type": "integer", "minimum": 0 },
          "successRate": { "type": "number", "minimum": 0, "maximum": 100 },
          "avgErrors": { "type": "number", "minimum": 0 },
          "avgHints": { "type": "number", "minimum": 0 },
          "avgRetries": { "type": "number", "minimum": 0 },
          "priorityScore": { "type": "number", "minimum": 0, "maximum": 100 },
          "reason": { "type": "string" }
        }
      },
      "AdaptiveProfileInsightsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "adaptivePreferences",
          "adaptiveProfile",
          "categoryRecommendations",
          "generatedAt",
          "attemptWindowSize"
        ],
        "properties": {
          "adaptivePreferences": { "$ref": "#/components/schemas/AdaptivePreferences" },
          "adaptiveProfile": { "$ref": "#/components/schemas/AdaptiveProfile" },
          "categoryRecommendations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdaptiveCategoryRecommendation" }
          },
          "generatedAt": { "type": "string", "format": "date-time" },
          "attemptWindowSize": { "type": "integer", "minimum": 0 }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 50 },
          "avatarId": { "type": "string" },
          "adaptivePreferences": { "$ref": "#/components/schemas/AdaptivePreferences" }
        }
      },
      "AuthTokenResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["token", "user"],
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/SerializedUser" }
        }
      },
      "ForgotPasswordResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "resetToken": { "type": "string" }
        }
      },
      "ResetPasswordResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["success", "token", "user"],
        "properties": {
          "success": { "type": "boolean" },
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/SerializedUser" }
        }
      },
...
    }
  }
}